+++ tests/testthat/test-71-pxjob-metamake.R
@@ -25,7 +25,6 @@ test_that("px file and pxmake(metamake(px file)) are equivalent", {
         # Improve when implementing #163
         stringr::str_subset("^VARIABLECODE.+", negate = TRUE) %>%
         stringr::str_subset("^VARIABLE-TYPE.+", negate = TRUE) %>%
-        stringr::str_subset("^VALUENOTE.+", negate = TRUE) %>%
         stringr::str_subset("^META-ID.+", negate = TRUE)
 
       if (table_name %in% c("no_timeval_or_codes2")) {